+++ FluentWoW/Controls/ToggleSwitch/ToggleSwitch.lua
@@ -27,10 +27,11 @@ function ToggleSwitchMixin:OnStateChanged(newState, prevState)
     local state = newState
     local isOn = self._isOn
 
+    local sr, sg, sb = T:GetColor("Color.Surface.Base")
     if state == "Disabled" then
         self.Track.BG:SetColorTexture(T:GetColor("Color.Surface.Stroke"))
         self.Thumb.Dot:SetColorTexture(T:GetColor("Color.Icon.Disabled"))
-        self.Thumb.Shadow:SetColorTexture(T:GetColor("Color.Surface.Base"), 0.7)
+        self.Thumb.Shadow:SetColorTexture(sr, sg, sb, 0.7)
         self.HeaderLabel:SetTextColor(T:GetColor("Color.Text.Disabled"))
         self.StateLabel:SetTextColor(T:GetColor("Color.Text.Disabled"))
         self:SetAlpha(T:GetNumber("Opacity.Disabled"))
@@ -48,7 +49,7 @@ function ToggleSwitchMixin:OnStateChanged(newState, prevState)
             self.Thumb.Dot:SetColorTexture(T:GetColor("Color.Icon.Default"))
             self.StateLabel:SetTextColor(T:GetColor("Color.Text.Secondary"))
         end
-        self.Thumb.Shadow:SetColorTexture(T:GetColor("Color.Surface.Base"), 0.7)
+        self.Thumb.Shadow:SetColorTexture(sr, sg, sb, 0.7)
     end
 
     self:_UpdateContentLabel()